Description
There are an amazing (no) bug with PHP versions 5.4 and under:
If client sends wrong POST-request with too big 'Content-length', the server gives the query to the PHP interpreter and it allocates 'Content-length' bytes in RAM, and only then check the query correctness. So if send 'Content-length' more than RAM size at server it will drop.